fix warning new since stdint.h change

+++ common/list.c
@@ -221,6 +221,6 @@ list_dump_items(struct list *self)
 
     for (index = 0; index < self->count; index++)
     {
-        g_writeln("%d: 0x%lx", index, list_get_item(self, index));
+        g_writeln("%d: 0x%lx", index, (long) list_get_item(self, index));
     }
 }
