+++ common/log.c
@@ -592,7 +592,7 @@ log_message(const enum logLevels lvl, const char *msg, ...)
         /* log to syslog*/
         /* %s fix compiler warning 'not a string literal' */
         syslog(internal_log_xrdp2syslog(lvl), "(%d)(%ld)%s", g_getpid(),
-               tc_get_threadid(), buff + 20);
+               (long) tc_get_threadid(), buff + 20);
     }
 
     if (lvl <= g_staticLogConfig->log_level)
